+++ oshi-json/src/main/java/oshi/json/hardware/Baseboard.java
@@ -1,0 +1,57 @@
+/**
+ * Oshi (https://github.com/dblock/oshi)
+ *
+ * Copyright (c) 2010 - 2016 The Oshi Project Team
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Maintainers:
+ * dblock[at]dblock[dot]org
+ * widdis[at]gmail[dot]com
+ * enrico.bianchi[at]gmail[dot]com
+ *
+ * Contributors:
+ * https://github.com/dblock/oshi/graphs/contributors
+ */
+package oshi.json.hardware;
+
+import oshi.json.json.OshiJsonObject;
+
+/**
+ * The Baseboard represents the system board, also called motherboard, logic
+ * board, etc.
+ * 
+ * @author widdis [at] gmail [dot] com
+ */
+public interface Baseboard extends OshiJsonObject {
+    /**
+     * Get the baseboard manufacturer.
+     * 
+     * @return The manufacturer.
+     */
+    String getManufacturer();
+
+    /**
+     * Get the baseboard model.
+     * 
+     * @return The model.
+     */
+    String getModel();
+
+    /**
+     * Get the baseboard version.
+     * 
+     * @return The version.
+     */
+    String getVersion();
+
+    /**
+     * Get the baseboard serial number
+     * 
+     * @return The serial number.
+     */
+    String getSerialNumber();
+}