+++ oshi-json/src/main/java/oshi/json/hardware/HardwareAbstractionLayer.java
@@ -29,6 +29,15 @@ import oshi.json.json.OshiJsonObject;
 public interface HardwareAbstractionLayer extends OshiJsonObject {
 
     /**
+     * Instantiates a {@link ComputerSystem} object. This represents the
+     * physical hardware, including components such as BIOS/Firmware and a
+     * motherboard, logic board, etc.
+     * 
+     * @return a {@link ComputerSystem} object.
+     */
+    ComputerSystem getComputerSystem();
+
+    /**
      * Instantiates a {@link CentralProcessor} object. This represents one or
      * more Logical CPUs.
      *